+++ packages/client/src/publish/Publisher.ts
@@ -14,8 +14,8 @@ import { Mapping } from '../utils/Mapping'
 
 export class PublishError extends Error {
 
-    public streamId: StreamID
-    public timestamp: number
+    public readonly streamId: StreamID
+    public readonly timestamp: number
 
     constructor(streamId: StreamID, timestamp: number, cause: Error) {
         // Currently Node and Firefox show the full error chain (this error and