+++ packages/client/src/subscribe/Resends.ts
@@ -1,5 +1,5 @@
 import { inject, Lifecycle, scoped, delay } from 'tsyringe'
-import { MessageRef, StreamPartID, StreamPartIDUtils, toStreamPartID } from '@streamr/protocol'
+import { StreamPartID, StreamPartIDUtils, toStreamPartID } from '@streamr/protocol'
 
 import { MessageStream } from './MessageStream'
 import { createSubscribePipeline } from './subscribePipeline'
@@ -27,7 +27,7 @@ const MIN_SEQUENCE_NUMBER_VALUE = 0
 
 type QueryDict = Record<string, string | number | boolean | null | undefined>
 
-export type ResendRef = MessageRef | {
+export interface ResendRef {
     timestamp: number | Date | string
     sequenceNumber?: number
 }