+++ packages/protocol/src/protocol/message_layer/MessageID.ts
@@ -4,6 +4,7 @@ import MessageRef from './MessageRef'
 import { StreamID, toStreamID } from '../../../src/utils/StreamID'
 import { StreamPartID, toStreamPartID } from "../../utils/StreamPartID"
 import { EthereumAddress, toEthereumAddress } from '@streamr/utils'
+/** @internal */
 export type MessageIDArray = [string, number, number, number, string, string]
 export default class MessageID {